refactor: split Cli into locator, download, and run files
