+++ editor/Editor/Cli.Download.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: adc41e9a1f27402a9ba6e79f98f1d0b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
