+++ editor/Editor/Cli.Run.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b45038072fff4f1689a25ab04c6cfc98
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
